Drop explicit doc of single-file mode
This mode is gone in the latest stable version (12) so it's a bit
confusing to describe the behavior completely. And it will only get less
useful over time.

Instead, just note that cross-file is supported. It's tagged with v11,
when it became the default. (You could opt in in v10 but it wasn't great)