Compilation Database: Point Bazel users to a solution
This doc lists ways of getting compilation databases out of some popular build systems for use with clangd. 

We built such a way for Bazel and just released it after a bunch of requests on GitHub. I'd [proposed adding it to the llvm compile_commands.json docs](https://reviews.llvm.org/D114213), and folks suggested we also add it here.

Thanks for your consideration--and the most wonderful of tools.